pyproject.toml
[tool.poetry]
name = "art-vostorg-store"
version = "0.1.0"
description = ""
authors = ["Mikhail <mikhailfedorov1939@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
django-oscar = "^3.2"
sorl-thumbnail = "^12.9.0"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"
django-filter = "^23.1"
dj-database-url = "^1.3.0"
boto3 = "^1.26.114"
django-storages = "^1.13.2"
pycountry = "^22.3.5"
django-redis = "^5.2.0"
django-cleanup = "^7.0.0"
django-oscar-api = "^3.2.0"
python-dotenv = "^1.0.0"
django-debug-toolbar = "^4.0.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
django-cors-headers = "^3.14.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"


[tool.black]
target-version = ['py310']
extend-exclude = '''(urls.py|migrations)'''

[tool.isort]
extend_skip = ["migrations"]
line_length = 88
profile = "black"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
